refactor(cli): lift orchestration into lib.rs and rename pipeline/ to stages/
The previous structure had two layers of orchestration indirection:
`lib.rs` was a thin facade that re-exported from `pipeline/mod.rs`,
which held `FileTask` / `FileOutcome` / `run` / `process_file`.
`pub use pipeline::*` added no value and made `cli::FileTask` physically
reside at `cli::pipeline::FileTask`, an internal detail.

Collapse the two layers:

- `lib.rs` is now the program itself: it owns `FileTask`,
  `FileOutcome`, `ProcessingReport`, `RunSummary`, `process_file`, and
  the six-stage composition. `run` is redefined in the TRPL style —
  it takes argv (`impl Iterator<Item = String>`) and returns
  `Result<RunSummary, CliError>`, folding what previously lived in
  `main.rs` (iteration, per-outcome logging, processed/error counting)
  into the library so it is testable.
- `pipeline/` is renamed to `stages/` and shrinks to a `mod.rs` with
  only submodule declarations. The folder now matches its content
  (stage implementations) rather than naming an activity that has
  moved elsewhere.
- `main.rs` becomes a 13-line OS shell: initialise `env_logger`, call
  `cli::run(env::args())`, translate the outcome to `ExitCode`.
- `for_testing::parse_and_structure` now routes through
  `crate::stages::*`.
- The single integration test that used `run(tasks)` is updated to use
  the new argv-based `cli::run(args)` and assert against
  `RunSummary`.

No behavior change; 62 tests pass (cli 31 / integration 6 / motorsport
25). clippy: no new warnings.
